refactored all modules, changed dto structure, fixed saving images logic

+++ src/comment/comments.service.ts
@@ -1,9 +1,12 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 
-import { ImageAction, CreateImageDto } from '../image/dto/images.dto';
-import { CreatePostCommentDto, UpdatePostCommentDto } from './dto/comments.dto';
+import {
+  CreatePostCommentDto,
+  UpdatePostCommentDto,
+  Action,
+} from './dto/comments.dto';
 
 import { CommentsEntity } from './comments.entity';
 
@@ -27,12 +30,18 @@ export class CommentsService {
     const postComment = this.postCommentsRepository.create(postCommentData);
 
     postComment.user = await this.usersService.getById(user_id);
+    if (!postComment.user) {
+      throw new NotFoundException('User with this id is not exist');
+    }
 
     postComment.post = await this.postsService.getById(post_id);
+    if (!postComment.post) {
+      throw new NotFoundException('Post with this id is not exist');
+    }
 
-    await Promise.all(
+    postComment.images = await Promise.all(
       images.map(async (image) => {
-        postComment.images.push(await this.imagesService.create(image));
+        return await this.imagesService.create(image);
       }),
     );
 
@@ -40,28 +49,22 @@ export class CommentsService {
   }
 
   async getAll(): Promise<CommentsEntity[]> {
-    return await this.postCommentsRepository.find({
-      relations: ['post', 'user', 'images'],
-    });
+    return await this.postCommentsRepository.find();
   }
 
   async getById(id: string): Promise<CommentsEntity> {
-    return await this.postCommentsRepository.findOne(id, {
-      relations: ['post', 'user', 'images'],
-    });
+    return await this.postCommentsRepository.findOne(id);
   }
 
   async getAllByUserId(id: string): Promise<CommentsEntity[]> {
     return await this.postCommentsRepository.find({
       where: { user: { id: id } },
-      relations: ['post', 'user', 'images'],
     });
   }
 
   async getAllByPostId(id: string): Promise<CommentsEntity[]> {
     return await this.postCommentsRepository.find({
       where: { post: { id: id } },
-      relations: ['post', 'user', 'images'],
     });
   }
 
@@ -69,18 +72,20 @@ export class CommentsService {
     const { image_actions, ...dataToUpdate } = data;
     const comment = await this.postCommentsRepository.findOne(id);
 
+    if (!comment) {
+      throw new NotFoundException('Comment with this id is not exist');
+    }
+
     if (image_actions.length !== 0) {
       await Promise.all(
         image_actions.map(async (image_action) => {
-          switch (image_action.action) {
-            case ImageAction.ADD:
-              const image = await this.imagesService.create(
-                image_action.data as CreateImageDto,
-              );
+          switch (image_action.type) {
+            case Action.ADD:
+              const image = await this.imagesService.create(image_action.data);
               comment.images.push(image);
               break;
-            case ImageAction.DELETE:
-              const id = image_action.data as string;
+            case Action.DELETE:
+              const id = image_action.id;
               comment.images = comment.images.filter((image) => {
                 return image.id !== id;
               });
@@ -96,12 +101,18 @@ export class CommentsService {
   }
 
   async remove(id: string): Promise<void> {
-    const { images } = await this.postCommentsRepository.findOne(id);
-    await Promise.all(
-      images.map(async (image) => {
-        await this.imagesService.remove(image.id);
-      }),
-    );
+    const comment = await this.postCommentsRepository.findOne(id);
+
+    if (!comment) {
+      throw new NotFoundException('Comment with this id is not exist');
+    }
+    if (comment.images.length !== 0) {
+      await Promise.all(
+        comment.images.map(async (image) => {
+          await this.imagesService.remove(image.id);
+        }),
+      );
+    }
     await this.postCommentsRepository.delete(id);
   }
 }
