+++ src/post/posts.service.ts
@@ -1,9 +1,8 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 
-import { ImageAction, CreateImageDto } from '../image/dto/images.dto';
-import { CreatePostDto, UpdatePostDto, CategoryAction } from './dto/posts.dto';
+import { CreatePostDto, UpdatePostDto, Action } from './dto/posts.dto';
 
 import { PostsEntity } from './posts.entity';
 
@@ -28,40 +27,42 @@ export class PostsService {
 
     post.user = await this.usersService.getById(user_id);
 
-    category_ids.map(async (categoryId) => {
-      post.categories.push(await this.categoriesService.getById(categoryId));
-    });
+    if (!post.user) {
+      throw new NotFoundException('User with this id is not exist');
+    }
 
-    images.map(async (image) => {
-      post.images.push(await this.imagesService.create(image));
-    });
+    post.categories = await Promise.all(
+      category_ids.map(async (categoryId) => {
+        return await this.categoriesService.getById(categoryId);
+      }),
+    );
+
+    post.images = await Promise.all(
+      images.map(async (image) => {
+        return await this.imagesService.create(image);
+      }),
+    );
 
     await this.postsRepository.save(post);
   }
 
   async getAll(): Promise<PostsEntity[]> {
-    return await this.postsRepository.find({
-      relations: ['categories', 'user', 'comments', 'images'],
-    });
+    return await this.postsRepository.find();
   }
 
   async getById(id: string): Promise<PostsEntity> {
-    return await this.postsRepository.findOne(id, {
-      relations: ['categories', 'user', 'comments', 'images'],
-    });
+    return await this.postsRepository.findOne(id);
   }
 
   async getAllByCategoryId(categoryId: string): Promise<PostsEntity[]> {
     return await this.postsRepository.find({
       where: { categories: { id: categoryId } },
-      relations: ['categories', 'user', 'comments', 'images'],
     });
   }
 
   async getAllByUserId(userId: string): Promise<PostsEntity[]> {
     return await this.postsRepository.find({
       where: { user: { id: userId } },
-      relations: ['categories', 'user', 'comments', 'images'],
     });
   }
 
@@ -72,13 +73,13 @@ export class PostsService {
     await Promise.all(
       category_actions.map(async (action) => {
         switch (action.type) {
-          case CategoryAction.ADD:
+          case Action.ADD:
             const category = await this.categoriesService.getById(
               action.category_id,
             );
             post.categories.push(category);
             break;
-          case CategoryAction.DELETE:
+          case Action.DELETE:
             post.categories = post.categories.filter((category) => {
               return category.id !== action.category_id;
             });
@@ -91,15 +92,13 @@ export class PostsService {
 
     await Promise.all(
       image_actions.map(async (image_action) => {
-        switch (image_action.action) {
-          case ImageAction.ADD:
-            const image = await this.imagesService.create(
-              image_action.data as CreateImageDto,
-            );
+        switch (image_action.type) {
+          case Action.ADD:
+            const image = await this.imagesService.create(image_action.data);
             post.images.push(image);
             break;
-          case ImageAction.DELETE:
-            const id = image_action.data as string;
+          case Action.DELETE:
+            const id = image_action.id;
             post.images = post.images.filter((image) => {
               return image.id !== id;
             });
@@ -115,12 +114,18 @@ export class PostsService {
   }
 
   async remove(id: string): Promise<void> {
-    const { images } = await this.postsRepository.findOne(id);
-    await Promise.all(
-      images.map(async (image) => {
-        await this.imagesService.remove(image.id);
-      }),
-    );
+    const post = await this.postsRepository.findOne(id);
+
+    if (!post) {
+      throw new NotFoundException('Post with this id is not exist');
+    }
+    if (post.images.length !== 0) {
+      await Promise.all(
+        post.images.map(async (image) => {
+          await this.imagesService.remove(image.id);
+        }),
+      );
+    }
     await this.postsRepository.delete(id);
   }
 }
